fix: page now loads with data when PAT is entered
