+++ src/lib/components/overview-page/Heading.svelte
@@ -3,10 +3,8 @@
     import { create_dropdown_selection } from "$lib/stores/dropdown";
     import Icon from "@iconify/svelte";
     import ButtonTintedMedium from "$lib/components/global/ButtonTintedMedium.svelte";
-    import AccessTokenModal from "../global/AccessTokenModal.svelte";
     import DropdownTintedMedium from "../global/DropdownTintedMedium.svelte";
     import Tab from "../global/Tab.svelte";
-    import { auth_error, retry_clone_with_token } from "$lib/stores/auth";
 
     let { repo_path: repo_path, repo_type: repo_type = "github" } = $props();
 
@@ -28,23 +26,6 @@
         selected_view = id;
     }
 
-    // Subscribe to auth errors to show modal when needed
-    let show_modal = $derived($auth_error.needs_token);
-
-    async function handle_token_add(token: string) {
-        console.log("Authenticating with Personal Access Token...");
-        
-        // Attempt to clone with the provided token
-        const success = await retry_clone_with_token(token);
-        
-        if (success) {
-            console.log("Authentication successful, refreshing page...");
-            // Refresh the page to load the new repository data
-            window.location.reload();
-        } else {
-            console.log("Authentication failed, please check your token");
-        }
-    }
 
     function open_calendar() {
         //calendar logic
@@ -75,15 +56,9 @@
                 label_class="body-accent"
                 icon_first={true}
                 width="4rem"
-				on:click={() => (show_modal = true)}
             />
         </div>
 
-            <!-- Access Token Modal -->
-            <AccessTokenModal 
-                bind:show_modal 
-                on_token_add={handle_token_add}
-            />
 
             <!-- branch dropdown btn -->
             <DropdownTintedMedium
